Workshop Page Fixes
Made the "events-header" texts more consistent in the Professional.js file. Had to remove the carousel portion of the professional workshops page in the Professional.js file in order to fix the posters stretching out and to make in more in line with the other posters on the page. The carousel did not work well with the "justify-content-md-center." Also, added a ".justify-content-md-center" to the professional.css file in order to get everything working well.