Added new js and css file containing Modal, made changes to Professional.js and professional.css, added the images we are practicing on
This contains the progress of what we have for the modal assignment. Added PictureModal.js and PictureModal.css containing all the modal functions. Professional.js and professional.css have been changed appropriately to display the contents of the "Event Gallery". Finally, it also contains the images we used for testing.